refactor: Update ComputersManager.vue to handle disconnection events
This commit updates the ComputersManager.vue component to handle disconnection events. When an event occurs, the code checks if there are any connected computers in the props.computers array. If there are, it sets the 'isConnected' property of the first connected computer to true and updates the selectedComputer value accordingly.